First working version of new-deps
