+++ cumulusci/core/dependencies/dependencies.py
@@ -1,24 +1,20 @@
-import abc
+from cumulusci.utils.yaml.model_parser import CCIModel
 import io
 import logging
-from typing import List, Optional
+from typing import ForwardRef, Iterable, List, Optional, Tuple
 
 import pydantic
 from github3.exceptions import NotFoundError
 from github3.repos.repo import Repository
-from pydantic import BaseModel
 from pydantic.networks import AnyUrl
 
 from cumulusci.core.config import OrgConfig
 from cumulusci.core.config.project_config import BaseProjectConfig
-from cumulusci.core.dependencies.resolvers import (
-    DependencyResolutionStrategy,
-    get_resolver,
-)
 from cumulusci.core.exceptions import DependencyResolutionError
 from cumulusci.salesforce_api.metadata import ApiDeploy
 from cumulusci.salesforce_api.package_install import (
     DEFAULT_PACKAGE_RETRY_OPTIONS,
+    ManagedPackageInstallOptions,
     install_1gp_package_version,
     install_package_version,
 )
@@ -26,14 +22,43 @@ from cumulusci.salesforce_api.package_zip import MetadataPackageZipBuilder
 from cumulusci.utils import download_extract_github_from_repo, download_extract_zip
 from cumulusci.utils.git import split_repo_url
 from cumulusci.utils.yaml.cumulusci_yml import cci_safe_load
+import abc
+
+from cumulusci.core.github import (
+    find_latest_release,
+    find_repo_feature_prefix,
+    get_version_id_from_commit,
+)
+from cumulusci.utils.git import (
+    get_feature_branch_name,
+    is_release_branch_or_child,
+    construct_release_branch_name,
+    get_release_identifier,
+)
+
+from enum import Enum
+
+from cumulusci.core.exceptions import CumulusCIException
+import itertools
+
 
 logger = logging.getLogger(__name__)
 
 
-class HashableBaseModel(BaseModel):
+class HashableBaseModel(CCIModel):
     # See https://github.com/samuelcolvin/pydantic/issues/1303
     def __hash__(self):
         return hash((type(self),) + tuple(self.__dict__.values()))
+
+
+class DependencyResolutionStrategy(str, Enum):
+    STRATEGY_STATIC_TAG_REFERENCE = "tag"
+    STRATEGY_2GP_EXACT_BRANCH = "exact_branch_2gp"
+    STRATEGY_2GP_RELEASE_BRANCH = "release_branch_2gp"
+    STRATEGY_2GP_PREVIOUS_RELEASE_BRANCH = "previous_release_branch_2gp"
+    STRATEGY_BETA_RELEASE_TAG = "latest_beta"
+    STRATEGY_RELEASE_TAG = "latest_release"
+    STRATEGY_UNMANAGED_HEAD = "unmanaged"
 
 
 class Dependency(HashableBaseModel, abc.ABC):
@@ -56,6 +81,9 @@ class Dependency(HashableBaseModel, abc.ABC):
         pass
 
 
+Dependency.update_forward_refs()
+
+
 class StaticDependency(Dependency, abc.ABC):
     @abc.abstractmethod
     def install(self, org_config: OrgConfig, retry_options: dict = None):
@@ -71,6 +99,8 @@ class StaticDependency(Dependency, abc.ABC):
 
 
 class DynamicDependency(Dependency, abc.ABC):
+    managed_dependency: Optional[StaticDependency]
+
     @property
     def is_flattened(self):
         return False
@@ -86,8 +116,12 @@ class DynamicDependency(Dependency, abc.ABC):
 
             if resolver and resolver.can_resolve(self, context):
                 try:
+                    context.logger.debug(f"Attempting to resolve {self} via {resolver}")
                     self.ref, self.managed_dependency = resolver.resolve(self, context)
                     if self.ref:
+                        context.logger.debug(
+                            f"Resolved {self} to {self.ref} (package {self.managed_dependency})"
+                        )
                         break
                 except DependencyResolutionError:
                     context.logger.info(
@@ -96,6 +130,23 @@ class DynamicDependency(Dependency, abc.ABC):
 
         if not self.ref:
             raise DependencyResolutionError(f"Unable to resolve dependency {self}")
+
+
+class Resolver(abc.ABC):
+    name = "Resolver"
+
+    @abc.abstractmethod
+    def can_resolve(self, dep: DynamicDependency, context: BaseProjectConfig) -> bool:
+        pass
+
+    @abc.abstractmethod
+    def resolve(
+        self, dep: DynamicDependency, context: BaseProjectConfig
+    ) -> Tuple[Optional[str], Optional["ManagedPackageDependency"]]:
+        pass
+
+    def __str__(self):
+        return self.name
 
 
 class GitHubRepoMixin:
@@ -140,6 +191,7 @@ class GitHubDynamicDependency(GitHubRepoMixin, DynamicDependency):
             logger.warning(
                 "The dependency keys `repo_owner` and `repo_name` are deprecated. Use `github` instead."
             )
+        return values
 
     @pydantic.root_validator
     def check_complete(cls, values):
@@ -167,6 +219,8 @@ class GitHubDynamicDependency(GitHubRepoMixin, DynamicDependency):
             values[
                 "github"
             ] = f"https://github.com/{values['repo_owner']}/{values['repo_name']}"
+
+        return values
 
     def _flatten_unpackaged(
         self,
@@ -290,12 +344,8 @@ class GitHubDynamicDependency(GitHubRepoMixin, DynamicDependency):
 
         return deps
 
-
-class ManagedPackageInstallOptions(HashableBaseModel):
-    activate_remote_site_settings: bool = True
-    name_conflict_resolution: str = "Block"
-    password: Optional[str]
-    security_type: str = "FULL"
+    def __str__(self):
+        return f"Dependency: {self.github}"
 
 
 class ManagedPackageDependency(StaticDependency):
@@ -327,6 +377,7 @@ class ManagedPackageDependency(StaticDependency):
         ) or "package_version_id" in values, (
             "Must specify `namespace` and `version`, or `package_version_id`"
         )
+        return values
 
     def install(
         self,
@@ -395,13 +446,15 @@ class UnmanagedDependency(GitHubRepoMixin, StaticDependency):
         ), "Must specify `zip_url`, or `github` and `ref`"
 
         # Populate the `github` and `repo_name, `repo_owner` properties if not already populated.
-        if not values.get("repo_name"):
+        if not values.get("repo_name") or not values.get("repo_owner"):
             values["repo_owner"], values["repo_name"] = split_repo_url(values["github"])
 
         if not values.get("github"):
             values[
                 "github"
             ] = f"https://github.com/{values['repo_owner']}/{values['repo_name']}"
+
+        return values
 
     def install(self, context: BaseProjectConfig, org: OrgConfig):
         zip_src = None
@@ -450,8 +503,8 @@ class UnmanagedDependency(GitHubRepoMixin, StaticDependency):
 def parse_dependency(dep_dict: dict) -> Optional[Dependency]:
     for dependency_class in [
         ManagedPackageDependency,
-        UnmanagedDependency,
         GitHubDynamicDependency,
+        UnmanagedDependency,
     ]:
         try:
             dep = dependency_class.parse_obj(dep_dict)
@@ -459,3 +512,371 @@ def parse_dependency(dep_dict: dict) -> Optional[Dependency]:
                 return dep
         except pydantic.ValidationError:
             pass
+
+
+## Resolvers
+
+
+class GitHubPackageDataMixin:
+    def _get_package_data(
+        self, repo: Repository, ref: str
+    ) -> Tuple[str, Optional[str]]:
+        contents = repo.file_contents("cumulusci.yml", ref=ref)
+        cumulusci_yml = cci_safe_load(io.StringIO(contents.decoded.decode("utf-8")))
+
+        # Get the namespace from the cumulusci.yml if set
+        package_config = cumulusci_yml.get("project", {}).get("package", {})
+        namespace = package_config.get("namespace")
+        package_name = (
+            package_config.get("name_managed")
+            or package_config.get("name")
+            or "Package"
+        )
+
+        return package_name, namespace
+
+
+class GitHubTagResolver(GitHubPackageDataMixin, Resolver):
+    name = "GitHub Tag Resolver"
+
+    def can_resolve(self, dep: DynamicDependency, context: BaseProjectConfig) -> bool:
+        return isinstance(dep, GitHubDynamicDependency) and dep.tag is not None
+
+    def resolve(
+        self, dep: GitHubDynamicDependency, context: BaseProjectConfig
+    ) -> Tuple[Optional[str], Optional[ManagedPackageDependency]]:
+        try:
+            # Find the github release corresponding to this tag.
+            repo = dep.get_repo(context)
+            release = repo.release_from_tag(dep.tag)
+            ref = repo.ref(f"tags/{release.tag_name}").object.sha
+            package_name, namespace = self._get_package_data(repo, ref)
+
+            if not dep.unmanaged and not namespace:
+                raise DependencyResolutionError(
+                    f"The tag {dep.tag} in {dep.github} does not identify a managed release"
+                )
+
+            if not dep.unmanaged:
+                return (
+                    repo.tag(ref).object.sha,
+                    ManagedPackageDependency(
+                        namespace=namespace,
+                        version=release.name,
+                        package_name=package_name,
+                    ),
+                )
+            else:
+                return repo.tag(ref).object.sha, None
+        except NotFoundError:
+            raise DependencyResolutionError(f"No release found for tag {dep.tag}")
+
+
+class GitHubReleaseTagResolver(GitHubPackageDataMixin, Resolver):
+    name = "GitHub Release Resolver"
+    include_beta = False
+
+    def can_resolve(self, dep: DynamicDependency, context: BaseProjectConfig) -> bool:
+        return isinstance(dep, GitHubDynamicDependency)
+
+    def resolve(
+        self, dep: GitHubDynamicDependency, context: BaseProjectConfig
+    ) -> Tuple[Optional[str], Optional[ManagedPackageDependency]]:
+        repo = dep.get_repo(context)
+        release = find_latest_release(repo, include_beta=self.include_beta)
+        if release:
+            ref = repo.tag(repo.ref(f"tags/{release.tag_name}").object.sha).object.sha
+            package_name, namespace = self._get_package_data(repo, ref)
+
+            return (
+                ref,
+                ManagedPackageDependency(
+                    namespace=namespace, version=release.name, package_name=package_name
+                ),
+            )
+
+        return (None, None)
+
+
+class GitHubBetaReleaseTagResolver(GitHubReleaseTagResolver):
+    name = "GitHub Release Resolver (Betas)"
+    include_beta = True
+
+
+class GitHubUnmanagedHeadResolver(Resolver):
+    name = "GitHub Unmanaged Resolver"
+
+    def can_resolve(self, dep: DynamicDependency, context: BaseProjectConfig) -> bool:
+        return isinstance(dep, GitHubDynamicDependency) and dep.tag is not None
+
+    def resolve(
+        self, dep: GitHubDynamicDependency, context: BaseProjectConfig
+    ) -> Tuple[Optional[str], Optional[ManagedPackageDependency]]:
+        repo = dep.get_repo(context)
+        return (repo.branch(repo.default_branch).commit.sha, None)
+
+
+def _locate_2gp_package_id(remote_repo, release_branch, context_2gp):
+    version_id = None
+    count = 0
+    commit = release_branch.commit
+    while version_id is None and count < 5:
+        version_id = get_version_id_from_commit(remote_repo, commit.sha, context_2gp)
+        if version_id:
+            break
+        count += 1
+        if commit.parents:
+            commit = remote_repo.commit(commit.parents[0]["sha"])
+        else:
+            break
+
+    return version_id, commit
+
+
+class GitHubReleaseBranchMixin:
+    def can_resolve(self, dep: DynamicDependency, context: BaseProjectConfig) -> bool:
+        return self.is_valid_repo_context(context) and isinstance(
+            dep, GitHubDynamicDependency
+        )
+
+    def get_release_id(self, context: BaseProjectConfig) -> int:
+        if not context.repo_branch or not context.project__git__prefix_feature:
+            raise DependencyResolutionError(
+                "Cannot get current branch or feature branch prefix"
+            )
+        release_id = get_release_identifier(
+            context.repo_branch, context.project__git__prefix_feature
+        )
+        if not release_id:
+            raise DependencyResolutionError("Cannot get current release identifier")
+
+        return int(release_id)
+
+    def is_valid_repo_context(self, context) -> bool:
+        return (
+            context.repo_branch
+            and context.project__git__prefix_feature
+            and is_release_branch_or_child(
+                context.repo_branch, context.project__git__prefix_feature
+            )
+        )
+
+
+class GitHubReleaseBranch2GPResolver(
+    GitHubPackageDataMixin, GitHubReleaseBranchMixin, Resolver
+):
+    name = "GitHub Release Branch 2GP Resolver"
+    branch_depth = 1
+
+    def resolve(
+        self, dep: GitHubDynamicDependency, context: BaseProjectConfig
+    ) -> Tuple[Optional[str], Optional[ManagedPackageDependency]]:
+
+        release_id = self.get_release_id(context)
+        repo = context.get_github_repo(dep.github)
+        if not repo:
+            raise DependencyResolutionError(
+                f"Unable to access GitHub repository for {dep.github}"
+            )
+
+        try:
+            remote_branch_prefix = find_repo_feature_prefix(repo)
+        except Exception:
+            context.logger.info(
+                f"Could not find feature branch prefix for {repo.clone_url}. Unable to resolve 2GP packages."
+            )
+            return (None, None)
+
+        # We will check at least the release branch corresponding to our release id.
+        # We may be configured to check backwards on release branches.
+        release_branch = None
+        for i in range(0, self.branch_depth + 1):
+            try:
+                remote_matching_branch = construct_release_branch_name(
+                    remote_branch_prefix, str(release_id - i)
+                )
+
+                release_branch = repo.branch(remote_matching_branch)
+            except NotFoundError:
+                pass
+
+        if release_branch:
+            version_id, commit = _locate_2gp_package_id(
+                repo,
+                release_branch,
+                context.project__git__2gp_context,  # TODO: we are assuming the projects use the same 2GP context
+            )
+
+            if version_id:
+                context.logger.info(
+                    f"Located 2GP package version {version_id} for release {release_id} on {repo.clone_url} at commit {release_branch.commit.sha}"
+                )
+
+                package_name, _ = self._get_package_data(repo, commit.sha)
+
+                return commit.sha, ManagedPackageDependency(
+                    version_id=version_id, package_name=package_name
+                )
+
+        context.logger.warn(
+            f"No 2GP package version located for release {release_id} on {repo.clone_url}."
+        )
+        return (None, None)
+
+
+class GitHubPreviousReleaseBranch2GPResolver(GitHubReleaseBranch2GPResolver):
+    name = "GitHub Previous Release Branch 2GP Resolver"
+    branch_depth = 3
+
+
+class GitHubReleaseBranchExactMatch2GPResolver(
+    GitHubPackageDataMixin, GitHubReleaseBranchMixin, Resolver
+):
+    name = "GitHub Exact-Match 2GP Resolver"
+
+    def resolve(
+        self, dep: GitHubDynamicDependency, context: BaseProjectConfig
+    ) -> Tuple[Optional[str], Optional[ManagedPackageDependency]]:
+        release_id = self.get_release_id(context)
+
+        repo = context.get_github_repo(dep.github)
+        if not repo:
+            raise DependencyResolutionError(f"Unable to access repository {dep.github}")
+
+        try:
+            remote_branch_prefix = find_repo_feature_prefix(repo)
+        except Exception:
+            context.logger.info(
+                f"Could not find feature branch prefix for {repo.clone_url}. Unable to resolve 2GP packages."
+            )
+            return (None, None)
+
+        # Attempt exact match
+        try:
+            branch = get_feature_branch_name(
+                context.repo_branch, context.project__git__prefix_feature
+            )
+            release_branch = repo.branch(f"{remote_branch_prefix}{branch}")
+        except Exception:
+            context.logger.info(f"Exact-match branch not found for {repo.clone_url}.")
+            return (None, None)
+
+        version_id, commit = _locate_2gp_package_id(
+            repo,
+            release_branch,
+            context.project__git__2gp_context,
+        )
+
+        if version_id:
+            context.logger.info(
+                f"Located 2GP package version {version_id} for release {release_id} on {repo.clone_url} at commit {release_branch.commit.sha}"
+            )
+
+            package_name, _ = self._get_package_data(repo, commit.sha)
+
+            return commit.sha, ManagedPackageDependency(
+                version_id=version_id, package_name=package_name
+            )
+
+        context.logger.warn(
+            f"No 2GP package version located for release {release_id} on {repo.clone_url}."
+        )
+        return (None, None)
+
+
+RESOLVER_CLASSES = {
+    DependencyResolutionStrategy.STRATEGY_STATIC_TAG_REFERENCE: GitHubTagResolver,
+    DependencyResolutionStrategy.STRATEGY_2GP_EXACT_BRANCH: GitHubReleaseBranchExactMatch2GPResolver,
+    DependencyResolutionStrategy.STRATEGY_2GP_RELEASE_BRANCH: GitHubReleaseBranch2GPResolver,
+    DependencyResolutionStrategy.STRATEGY_2GP_PREVIOUS_RELEASE_BRANCH: GitHubPreviousReleaseBranch2GPResolver,
+    DependencyResolutionStrategy.STRATEGY_BETA_RELEASE_TAG: GitHubBetaReleaseTagResolver,
+    DependencyResolutionStrategy.STRATEGY_RELEASE_TAG: GitHubReleaseTagResolver,
+    DependencyResolutionStrategy.STRATEGY_UNMANAGED_HEAD: GitHubUnmanagedHeadResolver,
+}
+
+
+## External API
+
+
+def get_resolver(
+    strategy: DependencyResolutionStrategy, dependency: DynamicDependency
+) -> Optional[Resolver]:
+    # This will be fleshed out when further types of DynamicDependency are added.
+
+    return RESOLVER_CLASSES[strategy]()
+
+
+def get_resolver_stack(
+    context: BaseProjectConfig, name: str
+) -> List[DependencyResolutionStrategy]:
+    resolutions = context.project__resolutions
+    stacks = context.project__resolutions__resolver_stacks
+
+    if name in resolutions and name != "resolver_stacks":
+        name = resolutions[name]
+
+    if stacks and name in stacks:
+        return [DependencyResolutionStrategy(n) for n in stacks[name]]
+
+    raise CumulusCIException(f"Resolver stack {name} was not found.")
+
+
+def get_static_dependencies(
+    dependencies: List[Dependency],
+    strategies: List[DependencyResolutionStrategy],
+    context: BaseProjectConfig,
+    ignore_deps: List[dict] = None,
+):
+    """Resolves the project -> dependencies section of cumulusci.yml
+    to convert dynamic github dependencies into static dependencies
+    by inspecting the referenced repositories.
+
+    Keyword arguments:
+    :param dependencies: a list of dependencies to resolve
+    :param ignore_deps: if provided, ignore the specified dependencies wherever found.
+    """
+
+    while any(not d.is_flattened or not d.is_resolved for d in dependencies):
+        for d in dependencies:
+            if not d.is_resolved:
+                d.resolve(context, strategies)
+
+        def unique(it: Iterable):
+            seen = set()
+
+            for each in it:
+                if each not in seen:
+                    seen.add(each)
+                    yield each
+
+        dependencies = list(
+            unique(
+                itertools.chain(
+                    *list(
+                        d.flatten(context)
+                        for d in dependencies
+                        if not _should_ignore_dependency(d, ignore_deps or [])
+                    )
+                ),
+            )
+        )
+
+    # Make sure, if we had no flattening or resolving to do, that we apply the ignore list.
+    return [
+        d for d in dependencies if not _should_ignore_dependency(d, ignore_deps or [])
+    ]
+
+
+def _should_ignore_dependency(dependency: Dependency, ignore_deps: List[dict]):
+    if not ignore_deps:
+        return False
+
+    ignore_github = [d["github"] for d in ignore_deps if "github" in d]
+    ignore_namespace = [d["namespace"] for d in ignore_deps if "namespace" in d]
+
+    if isinstance(dependency, ManagedPackageDependency) and dependency.namespace:
+        return dependency.namespace in ignore_namespace
+    if isinstance(dependency, GitHubDynamicDependency) and dependency.github:
+        return dependency.github in ignore_github
+
+    return False
